requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=1.5.0
pyarrow>=10.0.0
db-dtypes>=1.0.0
